requirements.txt
Pillow==5.0
